Work done on use cases of the allrecipes feature

+++ app/src/main/java/com/favorezapp/goodfood/allrecipes/domain/usecases/SearchAndGetRecipes.kt
@@ -1,7 +1,7 @@
 package com.favorezapp.goodfood.allrecipes.domain.usecases
 
 import com.favorezapp.goodfood.common.domain.RecipesRepository
-import com.favorezapp.goodfood.common.domain.model.NoMoreRecipesException
+import com.favorezapp.goodfood.common.domain.model.network.NetworkResponseState
 import com.favorezapp.goodfood.common.domain.model.pagination.PaginatedFoodRecipes
 import com.favorezapp.goodfood.common.domain.model.pagination.Pagination
 import javax.inject.Inject
@@ -12,16 +12,17 @@ class SearchAndGetRecipes @Inject constructor(
     suspend operator fun invoke(
         query: String,
         numOfItems: Int = Pagination.DEFAULT_NUM_OF_RECIPES
-    ): PaginatedFoodRecipes {
-        val (recipes, pagination) = repository.searchAndGetRecipes(numOfItems, query)
+    ): NetworkResponseState<PaginatedFoodRecipes> {
+        val state = repository.searchAndGetRecipes(numOfItems, query)
 
-        repository.deleteAllRecipes()
+        if (state is NetworkResponseState.Success<PaginatedFoodRecipes>) {
+            val (recipes, pagination) = state.result
+            if(recipes.isNotEmpty()) {
+                repository.deleteAllRecipes()
+                repository.storeFoodRecipes(recipes)
+            }
+        }
 
-        if( recipes.isNullOrEmpty() )
-            throw NoMoreRecipesException("No recipes found for $query")
-
-        repository.storeFoodRecipes( recipes )
-
-        return PaginatedFoodRecipes( recipes, pagination )
+        return state
     }
 }
